requirements.txt
aioredis==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
baron==0.10.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
debugpy==1.6.3
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
dj-rest-auth==2.2.4
Django==4.1
django-allauth==0.51.0
django-autoslug==1.9.8
django-cache-url==3.4.2
django-celery==3.1.17
django-channels==0.7.0
django-cities-light==3.9.0
django-compression-middleware==0.4.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-disqus==0.5
django-elasticsearch-dsl==7.2.2
django-environs==0.1.1
django-extensions==3.2.1
django-phonenumber-field==6.3.0
django-selenium-login==2.0.0
django-timezone-field==5.0
django-webpush==0.3.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
environs==9.5.0
h11==0.13.0
hiredis==2.0.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.16.0
mock==4.0.3
msgpack==1.0.4
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
persiantools==3.0.1
Pillow==9.1.1
pluggy==1.0.0
progressbar2==4.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptvsd==4.3.2
py==1.11.0
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-utils==3.3.3
python3-openid==3.2.0
pytz==2022.1
pywebpush==1.9.4
redbaron==0.9.2
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.4.4
rply==0.7.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.9
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.7.0
whitenoise==6.2.0
wsproto==1.1.0
zope.interface==5.4.0
zstandard==0.18.0
